Move `FeedViewModel`, `FeedLoadingViewModel` and `FeedErrorViewModel` to separate file Move  to separate file.
