+++ EssentialFeed/EssentialFeed/Feed Presentation/FeedPresenter.swift
@@ -7,27 +7,6 @@
 
 import Foundation
 
-public struct FeedErrorViewModel {
-    public let message: String?
-    
-    static var noError: FeedErrorViewModel {
-        return FeedErrorViewModel(message: nil)
-    }
-    
-    static func error(message: String) -> FeedErrorViewModel {
-        FeedErrorViewModel(message: message)
-    }
-}
-
-public struct FeedLoadingViewModel {
-    public let isLoading: Bool
-}
-
-public struct FeedViewModel {
-    public let feed: [FeedImage]
-}
-
-
 public protocol FeedView {
     func display(_ viewModel: FeedViewModel)
 }
