+++ EssentialFeed/EssentialFeed/Feed Presentation/FeedViewModel.swift
@@ -1,0 +1,12 @@
+//
+//  FeedViewModel.swift
+//  EssentialFeed
+//
+//  Created by Prabhat Tiwari on 18/07/25.
+//
+
+import Foundation
+
+public struct FeedViewModel {
+    public let feed: [FeedImage]
+}
